// src/util/interface.ts
export interface IPerson {
    name: string;
    lastName: string;
    country: string;
    birth: string;
    email: string;
    id: string
  }